+++ src/processor.js
@@ -10,22 +10,23 @@ class Processor {
   }
 
   datapackage (dataPackage) {
-    // todo: process schemas?
-    const resources = dataPackage.resources.map(resource => this.resource(resource));
-    return Object.assign(dataPackage, {resources});
+    dataPackage = Object.assign({}, dataPackage);
+    dataPackage.resources = dataPackage.resources.map(resource => this.resource(resource));
+    return dataPackage;
   }
 
   resource (resource) {
-    if (resource.content) {
-      const translator = this.translators[resource.mediatype];
+    const r = Object.assign({}, resource);
+    if (r.content) {
+      const translator = this.translators[r.mediatype];
       if (translator) {
-        Object.assign(resource, translator(resource));
+        Object.assign(r, translator(r));
       }
     }
-    if (resource.schema) {
-      Object.assign(resource, this.schemaProcessor.process(resource));
+    if (r.schema) {
+      return Object.assign(r, this.schemaProcessor.process(r));
     }
-    return resource;
+    return r;
   }
 }