+++ test/normalizer-package.js
@@ -10,10 +10,10 @@ nock('http://raw.githubusercontent.com')
   .get('/datasets/gdp/master/datapackage.json')
   .reply(200, gdp);
 
-test('normalize bare datapackage', async t => {
+test('normalize datapackage', async t => {
   const s = await dp.loadPackage('fixtures/bare/');
+  const p = dp.normalize.datapackage(s);
 
-  const p = dp.normalizeResources(dp.normalizePackage(s));
   t.not(p, s, 'creates a new object');
   t.not(p.resources, s.resources, 'creates a new resources array');
   t.regex(p.path, /fixtures\/bare\/$/, 'adds path');
@@ -25,8 +25,8 @@ test('normalize bare datapackage', async t => {
 
 test('normalize simple datapackage', async t => {
   const s = await dp.loadPackage('fixtures/inline/');
+  const p = dp.normalize.datapackage(s);
 
-  const p = dp.normalizeResources(dp.normalizePackage(s));
   t.not(p, s, 'creates a new object');
   t.not(p.resources, s.resources, 'creates a new resources array');
   t.regex(p.path, /fixtures\/inline\/$/, 'adds path');
@@ -41,8 +41,8 @@ test('normalize simple datapackage', async t => {
 
 test('normalize dpm generated datapackage', async t => {
   const s = await dp.loadPackage('fixtures/dpm/');
+  const p = dp.normalize.datapackage(s);
 
-  const p = dp.normalizeResources(dp.normalizePackage(s));
   t.not(p, s, 'creates a new object');
   t.not(p.resources, s.resources, 'creates a new resources array');
   t.regex(p.path, /fixtures\/dpm\/$/, 'adds path');
@@ -51,8 +51,8 @@ test('normalize dpm generated datapackage', async t => {
 
 test('normalize gdp datapackage', async t => {
   const s = await dp.loadPackage('fixtures/gdp/');
+  const p = dp.normalize.datapackage(s);
 
-  const p = dp.normalizeResources(dp.normalizePackage(s));
   t.not(s, p, 'creates a new object');
   t.not(p.resources, s.resources, 'creates a new resources array');
   t.regex(p.path, /fixtures\/gdp\/$/, 'adds path');
@@ -60,9 +60,9 @@ test('normalize gdp datapackage', async t => {
 });
 
 test('normalize gdp datapackage - from url', async t => {
-  const s = await dp.loadPackage('http://github.com/datasets/gdp');
+  const s = await dp.loader.datapackage('http://github.com/datasets/gdp');
+  const p = dp.normalize.datapackage(s);
 
-  const p = dp.normalizeResources(dp.normalizePackage(s));
   t.not(s, p, 'creates a new object');
   t.not(p.resources, s.resources, 'creates a new resources array');
   t.is(p.url, 'http://raw.githubusercontent.com/datasets/gdp/master/', 'adds path');